+++ R/core_api-future.R
@@ -160,14 +160,13 @@
 #' @aliases futureCall
 #' @rdname future
 #' @export
-future <- function(expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE, seed = FALSE, globals = TRUE, packages = NULL, stdout = TRUE, conditions = "condition", label = NULL, gc = FALSE, ...) {
+future <- function(expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE, seed = FALSE, globals = TRUE, packages = NULL, stdout = TRUE, conditions = "condition", label = NULL, ...) {
   debug <- isTRUE(getOption("future.debug"))
   if (debug) {
     mdebugf_push("future(..., label = %s) ...", sQuoteLabel(label))
     mdebugf("lazy: %s", lazy)
     mdebugf("stdout: %s", stdout)
     mdebugf("conditions: [n=%d] %s", length(conditions), commaq(conditions))
-    mdebugf("gc: %s", gc)
     on.exit(mdebugf_pop())
   }
   
@@ -198,14 +197,13 @@ future <- function(expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE
                    stdout = stdout,
                    conditions = conditions,
                    label = label,
-                   gc = gc,
                    onReference = onReference,
                    ...)
 
   args_names <- names(list(...))
   
-  ## Deprecation of arguments 'earlySignal'
-  for (name in c("earlySignal")) {
+  ## Deprecation of arguments 'earlySignal' and 'gc'
+  for (name in c("earlySignal", "gc")) {
     if (!is.null(future[[name]])) {
       if (is.element(name, args_names)) {
         deprecateFutureArgument(name, future[[name]])