+++ NEWS.md
@@ -22,9 +22,9 @@
 
 ## Deprecated and Defunct
 
- * Remove argument `earlySignal` from `future()`, `futureAssign()`,
-   and `futureCall()`.  Attempts to set it is now ignored and will
-   produce a deprecation warning.
+ * Remove arguments `earlySignal` and `gc` from `future()`,
+   `futureAssign()`, and `futureCall()`.  Attempts to set them are now
+   ignored and will produce deprecation warnings.
 
  * The `cluster` backend now defaults to `earlySignal = FALSE`. This
    was effectively already the case, because of an internal thinko